% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_classes.R
\docType{class}
\name{aligned_GCxGC-class}
\alias{aligned_GCxGC-class}
\title{Subclass aligned_GCxGC}
\description{
Subclass \emph{aligned_GCxGC} are contained in \emph{raw_GCxGC} super class.
It is not contained in the \emph{prepec_GCxGC} due to raw chromatograms can
be aligned without a previous preproccesing technique. Although, it can
improve the performance of the aligment, but it is not mandatory.
}
\details{
You can perform the aligment after some preprocessing technique as:
baseline correction, or signal smoothing to imporve the performance of the
alignment function.
}
