% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_classes.R
\docType{class}
\name{batch_2DCOW-class}
\alias{batch_2DCOW-class}
\title{Subclass batch_2DCOW}
\description{
Subclass \emph{batch_2DCOW} are contained in \emph{raw_GCxGC} super
class. \emph{batch_2DCOW} contains multiple aligned chromatograms, which the
first one is the reference chromatogram.
}
\details{
You can perform the alignment after some preprocessing technic as:
baseline correction, or signal smothing to improve the performance of the
aligment function, or with the raw chromatogram.
}
