% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_classes.R
\docType{class}
\name{foreign_load-class}
\alias{foreign_load-class}
\title{Class foreing_load}
\description{
Class \emph{forein_load} defines the superclass for the resulting loadings
 from foreing classification models, especially for PLS-DA model from
 mixOmics package
}
\section{Slots}{

\describe{
\item{\code{loadings}}{A matrix with the eigenvectors of each latent variable}

\item{\code{time}}{A two numeric vector of leght two with the range of the
chromatographic run time}

\item{\code{mod_time}}{The modulation period of the chromatographic experiment}

\item{\code{sampling_rate}}{The sampling rate of the mass analyzer}
}}

