% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\docType{methods}
\name{get_metadata}
\alias{get_metadata}
\alias{get_metadata,joined_chrom-method}
\alias{get_metadata,GCxGC-method}
\title{Method get_metadata}
\usage{
get_metadata(Object)

\S4method{get_metadata}{joined_chrom}(Object)
}
\arguments{
\item{chroms}{a joined chromatogram object created by
\code{\link{joined_chrom}} function.}
}
\description{
`get_metadata`retrieve the get_metadata contained in a joined_chrom
 object.
}
\details{
This function acceses to the \emph{groups} slot created by the
 \code{\link{join_chrom}} function. The \emph{Names} are the names of the
 chromatograms.
}
\examples{
data(Myrothecium)
myr_data <- get_metadata(all_chrom)
myr_data
}
