% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{get_metadata}
\alias{get_metadata}
\title{Extract metadata from a joinded_chrom}
\usage{
get_metadata(Object)
}
\arguments{
\item{Object}{a joined_chrom object}
}
\description{
`get_metadata` if a metadata was provided when join_chromatogram
funcion was used, the matadata will be retrieve from de object.
}
\examples{
GB08_fl <- system.file("extdata", "08GB.cdf", package = "RGCxGC")
GB09_fl <- system.file("extdata", "09GB.cdf", package = "RGCxGC")
GB08 <- read_chrom(GB08_fl, 5L)
GB09 <- read_chrom(GB09_fl, 5L)
metadata <- data.frame(Names = c("GB08", "GB09"),
                       Type = c("Control", "Treatment"))
join_metadata <- join_chromatograms(GB08, GB09, groups = metadata)
get_metadata(join_metadata)
}
