% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_plsda.R
\docType{methods}
\name{import_mixOmics}
\alias{import_mixOmics}
\alias{import_mixOmics,GCxGC-method}
\alias{import_mixOmics,PLSDA-method}
\title{Import mixOmics discriminant analysis}
\usage{
import_mixOmics(chromatogram, model, mod_time, time_range, sampling_rt)

\S4method{import_mixOmics}{GCxGC}(chromatogram, model, mod_time,
  time_range, sampling_rt)
}
\arguments{
\item{chromatogram}{a typicial GCxGC imported or preprocessed chromatogram.}

\item{model}{a partial least square discriminant analysis based model, built
by mixOmics package.}

\item{mod_time}{modulation time of the second dimension}

\item{time_range}{an atomic vector of two with the time range
of chromatographic run}

\item{sampling_rate}{the sampling rate of the equipment}
}
\description{
`import_mixOmics` transform a mixOmics discriminant analysis
  lodings/eigenvectors into a two-dimensional loading matrix
}
\details{
This functions takes a model built through
  \code{\link[mixOmics]{plsda}} and \code{\link[mixOmics]{splsda}}, then,
  access to loading values and transform each dimension loading into
  a two-dimensional matrix. This matrix represents a typcial GCxGC
  chromatogram. By default, user can provide a chromatogram where the
  requiered information will be retrieved from. On the other hand,
  user can also provide all the needed information
  (\emph{mod_time}, \emph{time_range}, \emph{sampling_rt})
  to fold eigenvectors into a GCxGC chromatogram.
}
