% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_classes.R
\docType{class}
\name{joined_chrom-class}
\alias{joined_chrom-class}
\title{Class joined_chrom}
\description{
Class \emph{joined_chrom} defines the superclass to gather single
chromatogram, as well as batch chromatograms into a single list, prior to
multiway principal compoment analysis or unfolding them.
}
\section{Slots}{

\describe{
\item{\code{chromatograms}}{a named list with all chromatograms.}

\item{\code{time}}{the time range of the chromatographic run}

\item{\code{groups}}{a data.frame containing the experiment metadata with
a column named as \emph{Names}.}

\item{\code{mod_time}}{modulation time of the second dimension}
}}

