% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_classes.R
\docType{class}
\name{MPCA-class}
\alias{MPCA-class}
\title{Subclass MPCA
Class \emph{MPCA} defines the class of Multiway Principal Component
Analysis}
\description{
Subclass MPCA
Class \emph{MPCA} defines the class of Multiway Principal Component
Analysis
}
\section{Slots}{

\describe{
\item{\code{scores}}{A matrix with the eigenvalues of projected chromatograms
into principal components space.}
}}

