% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loading.R
\docType{methods}
\name{plot_loading}
\alias{plot_loading}
\alias{plot_loading,projected-method}
\alias{plot_loading,MPCA-method}
\title{Plot two-dimensional MPCA loadings}
\usage{
plot_loading(Object, type = "n", pc = 1, thresh, ...)

\S4method{plot_loading}{projected}(Object, type = "b", pc = 1, thresh, ...)
}
\arguments{
\item{Object}{a MPCA object}

\item{type}{the value type of loadings, \emph{p} for positive,
\emph{n} for negative, and \emph{b} for negative and positive
loading values.}

\item{pc}{the principal component to plot.}

\item{thresh}{numerical value. A threshold to remove low loading values.}

\item{...}{Other parameters passes to \code{\link[graphics]{filled.contour}}
function.}
}
\description{
`plot_loading` plot the loading of the a MPCA object.
}
\details{
This function takes the loadings of MPCA and eval if a certain
  variable was removed previous compute de MPCA and it fills the removed
  variables with zero. Then, the loadings are plotted considering one
  principle component at a time as a two-dimensional chromatogram.
}
\examples{

library(colorRamps)
data(MTBLS579)
# MPCA with mean-centered and scaled data
MTBLS579_mpca <- m_prcomp(MTBLS579)
# Negative loadings of the first principal component
\donttest{
plot_loading(MTBLS579_mpca, type = "n", pc = 1,
             color.palette = matlab.like)
# Positive loadings of the first principal component
plot_loading(MTBLS579_mpca, type = "p", pc = 1,
             color.palette = matlab.like)
}
}
