% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_classes.R
\docType{class}
\name{PLSDA-class}
\alias{PLSDA-class}
\title{Subclass PLSDA}
\description{
Class \emph{PLSDA} defines the class to import foreign results of
partial least squares-discriminant analysis performed with
mixOmics package.
}
