% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_classes.R
\docType{class}
\name{preproc_GCxGC-class}
\alias{preproc_GCxGC-class}
\title{Subclass preproc_GCxGC}
\description{
Subclass \emph{preproc_GCxGC} are contained in \emph{raw_GCxGC} super class.
It contains a dedicated slot to storage the preprocessed two-dimensional
chromatogram.
}
\details{
After reading a two-dimensional chromatogram, you can perform serveral
preprocessing techniques such as smoothing or baseline correction.
This action will create an object of a preproc_GCxGC subclass.
}
