% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_chrom.R
\name{reference_chrom}
\alias{reference_chrom}
\title{Make reference chromatogram}
\usage{
reference_chrom(chromatograms)
}
\arguments{
\item{chromatograms}{a joined_chrom object.}
}
\description{
`reference_chrom` makes a reference chromatogram by calculating the mean
of multiple chromatograms.
}
\details{
The aim of this function is to create a consensus chromatogram to be used
as a reference one in the peak alignment process. In other words, the new
chromatogram will be used as a template and other chromatograms will be
aligned against it. This function overlap the provided chromatograms and
compute the mean for each pixel in order to create a reference 
two-dimensional chromatogram.
}
\examples{

# Read chromatogram 1
GB08_fl <- system.file("extdata", "08GB.cdf", package = "RGCxGC")
MTBLS08 <- read_chrom(GB08_fl, mod_time = 5)

# Read chromatogram 2
GB09_fl <- system.file("extdata", "09GB.cdf", package = "RGCxGC")
MTBLS09 <- read_chrom(GB09_fl, mod_time = 5)

# Join chromatograms
joined <-  join_chromatograms(MTBLS08, MTBLS09)
reference <- reference_chrom(joined)
plot(reference, main = "Reference chromaogram")
}
