% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\docType{methods}
\name{scores}
\alias{scores}
\alias{scores,MPCA-method}
\title{Method plot_scores}
\usage{
scores(Object)

\S4method{scores}{MPCA}(Object)
}
\arguments{
\item{Object}{a MPCA object}
}
\description{
`scores` exports the scores matrix of a MPCA
 object.
}
\details{
This function takes the whole MPCA object and retrieves the score
  matrix.
}
\examples{

data(MTBLS579)
# MPCA with mean-centered and scaled data
MTBLS579_mpca <- m_prcomp(MTBLS579, center = TRUE, scale = TRUE)
# Export scores matrix
scores(MTBLS579_mpca)
}
