% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{set_metadata}
\alias{set_metadata}
\title{Set the metadata for a joined_chrom}
\usage{
set_metadata(Object, metadata)
}
\arguments{
\item{Object}{a joined_chrom object}

\item{metadata}{a data.frame containing the metadata. It must have a column}
}
\description{
`set_metadata` fill metadata slot of a joined chrom.
}
\examples{

GB08_fl <- system.file("extdata", "08GB.cdf", package = "RGCxGC")
GB09_fl <- system.file("extdata", "09GB.cdf", package = "RGCxGC")
GB08 <- read_chrom(GB08_fl, 5L)
GB09 <- read_chrom(GB09_fl, 5L)
extra_info <- data.frame(Names = c("GB08", "GB09"),
                       Type = c("Control", "Treatment"))
join_chrom <- join_chromatograms(GB08, GB09)
join_metadata <- set_metadata(join_chrom, metadata = extra_info)
}
