% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dephase_chrom.R
\docType{methods}
\name{dephase_chrom}
\alias{dephase_chrom}
\alias{dephase_chrom,GCxGC-method}
\title{Method dephase_chrom}
\usage{
dephase_chrom(Object, rel_dephase)

\S4method{dephase_chrom}{GCxGC}(Object, rel_dephase)
}
\description{
`dephase_chrom`  move the retention time in the second dimension of a
 bidimensional chromatogram. This procedure is ussually applied in cases
 when peaks are near at the final of the second dimension of the 
 chromatogram. The dephasing is performing by splitting the chromagram
 with the relative value provided.
 
 @param Object a GCxGC class object
 @param rel_dephase a numeric value from 0 to 100 with the relative dephasing
   position.
 @export
 @examples 
 
 GB08_fl <- system.file("extdata", "08GB.cdf", package = "RGCxGC")
 GB08 <- read_chrom(GB08_fl, 5L)
 plot(GB08, nlevels = 150, color.palette = matlab.like,
      main = "No dephased chromatogram")
 GB08_d25 <- dephase_chrom(GB08, 25)
 plot(GB08_d25, nlevels = 150, color.palette = matlab.like,
      main = "25% dephased chromatogram")
}
