% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_classes.R
\docType{class}
\name{GCxGC-class}
\alias{GCxGC-class}
\title{Class GCxGC}
\description{
Class \emph{GCxGC} defines the superclass of bidimensional comprehensive
gas chromatography
}
\details{
The validity function evaluates if the provied file can be readed in a
netCDF format. The validation employs the function 
\code{\link[RNetCDF]{open.nc}} to check if the provided file inherits to
NetCDF.
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of a netCDF file to which the data will be retrieved}

\item{\code{mod_time}}{A integer with the modulation time for the second dimension.
Note the integer should be provide with an \emph{L} at the end of 
the number.}
}}

