% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\docType{methods}
\name{get_metadata}
\alias{get_metadata}
\alias{get_metadata,joined_chrom-method}
\alias{get_metadata,GCxGC-method}
\title{Method get_metadata}
\usage{
get_metadata(chroms)

\S4method{get_metadata}{joined_chrom}(chroms)
}
\arguments{
\item{chroms}{a joined_chrom object created by
joined_chrom function.}
}
\description{
`get_metadata` retrieves the metadata contained in a
joined_chrom object.
}
\details{
This function accesses to the \emph{groups} slot created by the
 joined_chrom function. The \emph{Names} are the names of the
 chromatograms.
}
\examples{
data(Myrothecium)
myr_data <- get_metadata(Myrothecium)
myr_data
}
