% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_classes.R
\docType{class}
\name{joined_chrom-class}
\alias{joined_chrom-class}
\title{Class joined_chrom}
\description{
Class \emph{joined_chrom} defines the superclass to gather single
chromatogram as well batch cromatograms into a singe list previos
multiway principal compoment analysis
}
\section{Slots}{

\describe{
\item{\code{chromatograms}}{A named list with all chromatograms.}

\item{\code{groups}}{A data.frame with the metadata. A column \emph{Name} with
the same name of the chromatograms}

\item{\code{time}}{The time range of chromatographic run}

\item{\code{mod_time}}{modulation time of the second dimension}
}}

