% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_classes.R
\docType{class}
\name{preproc_GCxGC-class}
\alias{preproc_GCxGC-class}
\title{Subclass preproc_GCxGC}
\description{
Subclass \emph{preproc_GCxGC} are contained in \emph{raw_GCxGC} super class.
It contains a dedicated slot to storage the preprocessed bidimensional
chromatogram.
}
\details{
After reading a bidimensional chromatogram, you can perform serveral
 preprocessing technicas as smothing, or baseline correction. It will
create an object of subclass preproc_GCxGC.
}
