% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_classes.R
\docType{class}
\name{raw_GCxGC-class}
\alias{raw_GCxGC-class}
\title{Subclass raw_GCxGC}
\description{
Subclass \emph{raw_GCxGC} are contained in \emph{GCxGC} super class. It
contains a dedicated slot to storage the folded bidimensional chromatogram.
}
\details{
In the first creation of a \emph{raw_GCxGC} object, the slot for the
chromatogram. To read and fold the chromatogram use the function
\code{\link{read_chrom}}.
}
\section{Slots}{

\describe{
\item{\code{chromatogram}}{a numeric matrix.}

\item{\code{time}}{a vector of to elements with the range of the first dimenstion
retention time}
}}

