% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_chrom.R
\name{unfold_chrom}
\alias{unfold_chrom}
\title{Unfold two-dimensional chromatograms}
\usage{
unfold_chrom(Object)
}
\arguments{
\item{Object}{a batch_2DCOW or joined_chrom objects.}
}
\description{
`unfold` converts two-dimensional chromatograms into a one dimensional
vector. Then, all chromatograms are joined into a matrix.
}
\details{
This function takes a single argument, batch_2DCOW or joined_chrom objects
and extracts each chromatogram (matix) and it is unfolded to a one dimensional
vector. Then, each one dimensional vector is joined in a single matrix. The
column matrix represents a given retention time, while row represents samples.
Also, in order to keep information about chromatographic runs, the retention
times for both dimensions are also exported.
}
\examples{
 

GB08_fl <- system.file("extdata", "08GB.cdf", package = "RGCxGC")
GB09_fl <- system.file("extdata", "09GB.cdf", package = "RGCxGC")
GB08 <- read_chrom(GB08_fl, 5L)
GB09 <- read_chrom(GB09_fl, 5L)
join_gc <- join_chromatograms(GB08, GB09)
chromatograms_1D <- unfold_chrom(join_gc)
chrom_mt <- chromatograms_1D$chromatogram
chromatograms_1D$mod_time
chromatograms_1D$time
}
